S03E03

+++ agenda/app/src/main/java/br/com/alura/agenda/dao/AlunoDAO.java
@@ -20,7 +20,7 @@ public class AlunoDAO extends SQLiteOpenHelper
 {
     public AlunoDAO(Context context)
     {
-        super(context, "Agenda", null, 5);
+        super(context, "Agenda", null, 6);
     }
 
     @Override
@@ -34,7 +34,8 @@ public class AlunoDAO extends SQLiteOpenHelper
                 "site TEXT, " +
                 "nota REAL, " +
                 "caminhoFoto TEXT," +
-                "sincronizado INT DEFAULT 0);";
+                "sincronizado INT DEFAULT 0," +
+                "desativado INT DEFAULT 0);";
 
         db.execSQL(sql);
     }
@@ -84,7 +85,9 @@ public class AlunoDAO extends SQLiteOpenHelper
                 sql = "ALTER TABLE Alunos ADD COLUMN sincronizado INT"
                         + "DEFAULT 0";
                 db.execSQL(sql);
-
+            case 5:
+                sql = "ALTAR TABLE Alunos ADD COLUMN desativado INT DEFAULT 0";
+                db.execSQL(sql);
         }
 
     }
@@ -117,12 +120,13 @@ public class AlunoDAO extends SQLiteOpenHelper
         dados.put("nota", aluno.getNota());
         dados.put("caminhoFoto", aluno.getCaminhoFoto());
         dados.put("sincronizado", aluno.getSincronizado());
+        dados.put("desativado", aluno.getDesativado());
         return dados;
     }
 
     public List<Aluno> buscaAlunos()
     {
-        String sql = "SELECT * FROM Alunos;";
+        String sql = "SELECT * FROM Alunos WHERE desativado = 0;";
         SQLiteDatabase db = getReadableDatabase();
         Cursor c = db.rawQuery(sql, null);
 
@@ -147,7 +151,7 @@ public class AlunoDAO extends SQLiteOpenHelper
             aluno.setNota(c.getDouble(c.getColumnIndex("nota")));
             aluno.setCaminhoFoto(c.getString(c.getColumnIndex("caminhoFoto")));
             aluno.setSincronizado(c.getInt(c.getColumnIndex("sincronizado")));
-
+            aluno.setDesativado(c.getInt(c.getColumnIndex("desativado")));
             alunos.add(aluno);
         }
         return alunos;
@@ -158,7 +162,14 @@ public class AlunoDAO extends SQLiteOpenHelper
         SQLiteDatabase db = getWritableDatabase();
 
         String[] params = {aluno.getId()};
-        db.delete("Alunos", "id=?", params);
+        if(aluno.estaDesativado())
+        {
+            db.delete("Alunos", "id=?", params);
+        }else
+        {
+            aluno.desativa();
+            altera(aluno);
+        }
     }
 
     public void altera(Aluno aluno)
